requirements.txt
py-cord==2.7.1
python-dotenv==1.2.2